+++ lib/listener_token.dart
@@ -3,6 +3,7 @@ import 'package:uuid/uuid.dart';
 class ListenerToken {
   final tokenId = Uuid().v1();
 
+  /// Listener token returned when adding a change listener. The token is used for removing the added change listener.
   ListenerToken();
 
   Map<String, dynamic> toJson() => {"token": tokenId};