+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 ## 2.5.1
 
 * Updated documentation
+
+## 2.5.1+1
+
+* Changed return type of class Database documentWithId from Map to Document
+* Populated doc/api using dartdoc
+* Fixed some format issues